major cleanup/refactoring
