+++ src/commands/musiccontrol/pause.js
@@ -1,5 +1,5 @@
-const {ButtonBuilder, ButtonStyle, ActionRowBuilder} = require("discord.js");
-const {button} = require("../../utils/constants");
+const editQueueDashboard = require("../../utils/editQueueDashboard");
+
 module.exports = {
 
     callback: async (client, interaction) => {
@@ -16,19 +16,8 @@ module.exports = {
         await interaction.editReply('The current song has been paused.')
 
         if (queue.dashboard) {
-            if (queue.dashboard) {
-                let playing = queue.node.isPaused();
-                let loop = queue.repeatMode === 1;
-
-                const playPauseButton = new ButtonBuilder().setCustomId('Playing-PlayPause').setEmoji(playing ? button.play : button.pause).setStyle(playing ? ButtonStyle.Success : ButtonStyle.Secondary);
-                const skipButton = new ButtonBuilder().setCustomId('Playing-Skip').setEmoji(button.skip).setStyle(ButtonStyle.Secondary);
-                const stopButton = new ButtonBuilder().setCustomId('Playing-Stop').setEmoji(button.stop).setStyle(ButtonStyle.Danger);
-                const loopButton = new ButtonBuilder().setCustomId('Playing-Loop').setEmoji(button.loop).setStyle(loop ? ButtonStyle.Danger : ButtonStyle.Secondary);
-                const shuffleButton = new ButtonBuilder().setCustomId('Playing-Shuffle').setEmoji(button.shuffle).setStyle(ButtonStyle.Secondary);
-                const row = new ActionRowBuilder().addComponents(playPauseButton, skipButton, stopButton, loopButton, shuffleButton);
-
-                queue.dashboard.edit({ components: [row] });
-            }
+            const row = editQueueDashboard(queue);
+            queue.dashboard.edit({ components: [row] });
         }
     },
