+++ src/commands/musiccontrol/toggleloop.js
@@ -1,5 +1,5 @@
-const {ButtonBuilder, ButtonStyle, ActionRowBuilder} = require("discord.js");
-const {button} = require("../../utils/constants");
+const editQueueDashboard = require("../../utils/editQueueDashboard");
+
 module.exports = {
 
     callback: async (client, interaction) => {
@@ -11,9 +11,7 @@ module.exports = {
             return interaction.editReply({content: 'There is no song playing.', ephemeral: true});
         }
 
-        const loop = queue.repeatMode === 1;
-
-        if (loop) {
+        if (queue.repeatMode === 1) {
             queue.setRepeatMode(0);
             await interaction.editReply('Looping has been disabled.')
         } else {
@@ -23,15 +21,7 @@ module.exports = {
         }
 
         if (queue.dashboard) {
-            let playing = queue.node.isPaused();
-
-            const playPauseButton = new ButtonBuilder().setCustomId('Playing-PlayPause').setEmoji(playing ? button.play : button.pause).setStyle(playing ? ButtonStyle.Success : ButtonStyle.Secondary);
-            const skipButton = new ButtonBuilder().setCustomId('Playing-Skip').setEmoji(button.skip).setStyle(ButtonStyle.Secondary);
-            const stopButton = new ButtonBuilder().setCustomId('Playing-Stop').setEmoji(button.stop).setStyle(ButtonStyle.Danger);
-            const loopButton = new ButtonBuilder().setCustomId('Playing-Loop').setEmoji(button.loop).setStyle(!loop ? ButtonStyle.Danger : ButtonStyle.Secondary);
-            const shuffleButton = new ButtonBuilder().setCustomId('Playing-Shuffle').setEmoji(button.shuffle).setStyle(ButtonStyle.Secondary);
-            const row = new ActionRowBuilder().addComponents(playPauseButton, skipButton, stopButton, loopButton, shuffleButton);
-
+            const row = editQueueDashboard(queue);
             queue.dashboard.edit({ components: [row] });
         }
     },
