+++ src/events/discord-player/player.js
@@ -1,6 +1,7 @@
 const { EmbedBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
 
 const { button } = require('../../utils/constants');
+const editQueueDashboard = require("../../utils/editQueueDashboard");
 
 const registerPlayerEvents = (player, client) => {
     player.events.on('connection', async (queue) => {
@@ -14,19 +15,10 @@ const registerPlayerEvents = (player, client) => {
         const startEmbed = new EmbedBuilder().setTitle('Dashboard').setDescription('No music is currently playing.');
 
         queue.dashboard = await queue.metadata.channel.send({ embeds: [startEmbed], components: [row] });
-        return;
     });
 
     player.events.on('playerStart', async (queue, track) => {
-        const playing = queue.node.isPaused();
-        const loop = queue.repeatMode === 1;
-
-        const playPauseButton = new ButtonBuilder().setCustomId('Playing-PlayPause').setEmoji(playing ? button.play : button.pause).setStyle(ButtonStyle.Secondary);
-        const skipButton = new ButtonBuilder().setCustomId('Playing-Skip').setEmoji(button.skip).setStyle(ButtonStyle.Secondary);
-        const stopButton = new ButtonBuilder().setCustomId('Playing-Stop').setEmoji(button.stop).setStyle(ButtonStyle.Danger);
-        const loopButton = new ButtonBuilder().setCustomId('Playing-Loop').setEmoji(button.loop).setStyle(loop ? ButtonStyle.Danger : ButtonStyle.Secondary);
-        const shuffleButton = new ButtonBuilder().setCustomId('Playing-Shuffle').setEmoji(button.shuffle).setStyle(ButtonStyle.Secondary);
-        const row = new ActionRowBuilder().addComponents(playPauseButton, skipButton, stopButton, loopButton, shuffleButton);
+        const row = editQueueDashboard(queue);
 
         const dashboardEmbed = new EmbedBuilder()
             .setTitle('Dashboard')
