+++ src/events/interactionCreate/handleButtons.js
@@ -2,6 +2,7 @@ const { ActionRowBuilder, ButtonBuilder, ButtonStyle, EmbedBuilder} = require('d
 
 const { wait } = require('../../utils/wait');
 const { button } = require('../../utils/constants');
+const editQueueDashboard = require("../../utils/editQueueDashboard");
 
 module.exports = async (client, interaction) => {
     if (!interaction.isButton()) return;
@@ -23,22 +24,13 @@ module.exports = async (client, interaction) => {
 
     switch (interaction.customId) {
         case 'Playing-PlayPause': {
-            let playing = !queue.node.isPaused();
-
-            if (playing) {
-                queue.node.pause();
-            } else {
+            if (queue.node.isPaused()) {
                 queue.node.resume();
+            } else {
+                queue.node.pause();
             }
 
-            const loop = queue.repeatMode === 1;
-
-            const playPauseButton = new ButtonBuilder().setCustomId('Playing-PlayPause').setEmoji(playing ? button.play : button.pause).setStyle(playing ? ButtonStyle.Success : ButtonStyle.Secondary);
-            const skipButton = new ButtonBuilder().setCustomId('Playing-Skip').setEmoji(button.skip).setStyle(ButtonStyle.Secondary);
-            const stopButton = new ButtonBuilder().setCustomId('Playing-Stop').setEmoji(button.stop).setStyle(ButtonStyle.Danger);
-            const loopButton = new ButtonBuilder().setCustomId('Playing-Loop').setEmoji(button.loop).setStyle(loop ? ButtonStyle.Danger : ButtonStyle.Secondary);
-            const shuffleButton = new ButtonBuilder().setCustomId('Playing-Shuffle').setEmoji(button.shuffle).setStyle(ButtonStyle.Secondary);
-            const row = new ActionRowBuilder().addComponents(playPauseButton, skipButton, stopButton, loopButton, shuffleButton);
+            const row = editQueueDashboard(queue);
 
             await interaction.update({ components: [row] });
         } break;
@@ -53,34 +45,18 @@ module.exports = async (client, interaction) => {
                 await wait(500);
             }
 
-            const playing = queue.node.isPaused();
-            const loop = queue.repeatMode === 1;
-
-            const playPauseButton = new ButtonBuilder().setCustomId('Playing-PlayPause').setEmoji(playing ? button.play : button.pause).setStyle(playing ? ButtonStyle.Success : ButtonStyle.Secondary);
-            const skipButton = new ButtonBuilder().setCustomId('Playing-Skip').setEmoji(button.skip).setStyle(ButtonStyle.Secondary);
-            const stopButton = new ButtonBuilder().setCustomId('Playing-Stop').setEmoji(button.stop).setStyle(ButtonStyle.Danger);
-            const loopButton = new ButtonBuilder().setCustomId('Playing-Loop').setEmoji(button.loop).setStyle(loop ? ButtonStyle.Danger : ButtonStyle.Secondary);
-            const shuffleButton = new ButtonBuilder().setCustomId('Playing-Shuffle').setEmoji(button.shuffle).setStyle(ButtonStyle.Secondary);
-            const row = new ActionRowBuilder().addComponents(playPauseButton, skipButton, stopButton, loopButton, shuffleButton);
+            const row = editQueueDashboard(queue);
 
             await interaction.update({ components: [row] });
         } break;
         case 'Playing-Loop': {
-            const loop = queue.repeatMode === 1;
-            if (loop) {
+            if (queue.repeatMode === 1) {
                 queue.setRepeatMode(0);
             } else {
                 queue.setRepeatMode(1);
             }
 
-            let playing = queue.node.isPaused();
-
-            const playPauseButton = new ButtonBuilder().setCustomId('Playing-PlayPause').setEmoji(playing ? button.play : button.pause).setStyle(playing ? ButtonStyle.Success : ButtonStyle.Secondary);
-            const skipButton = new ButtonBuilder().setCustomId('Playing-Skip').setEmoji(button.skip).setStyle(ButtonStyle.Secondary);
-            const stopButton = new ButtonBuilder().setCustomId('Playing-Stop').setEmoji(button.stop).setStyle(ButtonStyle.Danger);
-            const loopButton = new ButtonBuilder().setCustomId('Playing-Loop').setEmoji(button.loop).setStyle(!loop ? ButtonStyle.Danger : ButtonStyle.Secondary);
-            const shuffleButton = new ButtonBuilder().setCustomId('Playing-Shuffle').setEmoji(button.shuffle).setStyle(ButtonStyle.Secondary);
-            const row = new ActionRowBuilder().addComponents(playPauseButton, skipButton, stopButton, loopButton, shuffleButton);
+            const row = editQueueDashboard(queue);
 
             await interaction.update({ components: [row] });
         } break;
